[CAMEL-11257] Moved HTTP connection class to io package.
