+++ components/camel-as2/camel-as2-api/src/main/java/org/apache/camel/component/as2/api/io/AS2BHttpServerConnection.java
@@ -1,4 +1,4 @@
-package org.apache.camel.component.as2.api;
+package org.apache.camel.component.as2.api.io;
 
 import java.io.IOException;
 import java.nio.charset.CharsetDecoder;
